refactored reblogQueries to repostQueries
