+++ src/providers/queries/postQueries/index.ts
@@ -1,7 +1,7 @@
 import * as postQueries from './postQueries';
 import * as wavesQueries from './wavesQueries';
 import * as pollQueries from './pollQueries';
-import * as reblogQueries from './reblogQueries';
+import * as repostQueries from './repostQueries';
 import * as feedQueries from './feedQueries';
 
-export { postQueries, wavesQueries, pollQueries, reblogQueries, feedQueries };
+export { postQueries, wavesQueries, pollQueries, repostQueries, feedQueries };
